requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-sendgrid-v5==1.2.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==21.2.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2025.1
